Enum rekordcrate::anlz::Content

pub enum Content {
    BeatGrid {
        unknown1: u32,
        unknown2: u32,
        beats: Vec<Beat>,
    },
    CueList {
        list_type: CueListType,
        unknown: u16,
        memory_count: u32,
        cues: Vec<Cue>,
    },
    ExtendedCueList {
        list_type: CueListType,
        cues: Vec<ExtendedCue>,
    },
    Path(String),
    VBR {
        unknown1: u32,
        unknown2: Vec<u8>,
    },
    WaveformPreview {
        len_preview: u32,
        unknown: u32,
        data: Vec<WaveformPreviewColumn>,
    },
    TinyWaveformPreview {
        len_preview: u32,
        unknown: u32,
        data: Vec<TinyWaveformPreviewColumn>,
    },
    WaveformDetail {
        len_entry_bytes: u32,
        len_entries: u32,
        unknown: u32,
        data: Vec<WaveformPreviewColumn>,
    },
    WaveformColorPreview {
        len_entry_bytes: u32,
        len_entries: u32,
        unknown: u32,
        data: Vec<WaveformColorPreviewColumn>,
    },
    WaveformColorDetail {
        len_entry_bytes: u32,
        len_entries: u32,
        unknown: u32,
        data: Vec<WaveformColorDetailColumn>,
    },
    SongStructure {
        len_entry_bytes: u32,
        len_entries: u16,
        mood: Mood,
        unknown1: u32,
        unknown2: u16,
        end_beat: u16,
        unknown3: u16,
        bank: Bank,
        unknown4: u8,
        data: Vec<Phrase>,
    },
    Unknown {
        header_data: Vec<u8>,
        content_data: Vec<u8>,
    },
}

Section content which differs depending on the section type.

Variants

BeatGrid

Fields

unknown1: u32

Unknown field.

unknown2: u32

Unknown field.

According to @flesniak, this is always 00800000.

beats: Vec<Beat>

Beats in this beatgrid.

All beats in the track.

CueList

Fields

list_type: CueListType

The types of cues (memory or hot) that this list contains.

unknown: u16

Unknown field

memory_count: u32

Unknown field.

cues: Vec<Cue>

Cues

List of cue points or loops (either hot cues or memory cues).

ExtendedCueList

Fields

list_type: CueListType

The types of cues (memory or hot) that this list contains.

cues: Vec<ExtendedCue>

Cues

List of cue points or loops (either hot cues or memory cues, extended version).

Variation of the original CueList that also adds support for more metadata such as comments and colors. Introduces with the Nexus 2 series players.

Path(String)

Tuple Fields

0: String

Path of the audio file that this analysis belongs to.

VBR

Fields

unknown1: u32

Unknown field.

unknown2: Vec<u8>

Unknown data.

Seek information for variable bitrate files (probably).

WaveformPreview

Fields

len_preview: u32

Unknown field.

unknown: u32

Unknown field (apparently always 0x00100000)

data: Vec<WaveformPreviewColumn>

Waveform preview column data.

Fixed-width monochrome preview of the track waveform.

TinyWaveformPreview

Fields

len_preview: u32

Unknown field.

unknown: u32

Unknown field (apparently always 0x00100000)

data: Vec<TinyWaveformPreviewColumn>

Waveform preview column data.

Smaller version of the fixed-width monochrome preview of the track waveform.

WaveformDetail

Fields

len_entry_bytes: u32

Size of a single entry, always 1.

len_entries: u32

Number of entries in this section.

unknown: u32

Unknown field (apparently always 0x00960000)

data: Vec<WaveformPreviewColumn>

Waveform preview column data.

Each entry represents one half-frame of audio data, and there are 75 frames per second, so for each second of track audio there are 150 waveform detail entries.

Variable-width large monochrome version of the track waveform.

Used in .EXT files.

WaveformColorPreview

Fields

len_entry_bytes: u32

Size of a single entry, always 1.

len_entries: u32

Number of entries in this section.

unknown: u32

Unknown field.

data: Vec<WaveformColorPreviewColumn>

Waveform preview column data.

Each entry represents one half-frame of audio data, and there are 75 frames per second, so for each second of track audio there are 150 waveform detail entries.

Variable-width large monochrome version of the track waveform.

Used in .EXT files.

WaveformColorDetail

Fields

len_entry_bytes: u32

Size of a single entry, always 1.

len_entries: u32

Number of entries in this section.

unknown: u32

Unknown field.

data: Vec<WaveformColorDetailColumn>

Waveform detail column data.

Variable-width large colored version of the track waveform.

Used in .EXT files.

SongStructure

Fields

len_entry_bytes: u32

Size of a single entry, always 1.

len_entries: u16

Number of entries in this section.

mood: Mood

Overall type of phrase structure.

unknown1: u32

Unknown field.

unknown2: u16

Unknown field.

end_beat: u16

Number of the beat at which the last recognized phrase ends.

unknown3: u16

Unknown field.

bank: Bank

Stylistic bank assigned in Lightning Mode.

unknown4: u8

Unknown field.

data: Vec<Phrase>

Phrase entry data.

Describes the structure of a sond (Intro, Chrous, Verse, etc.).

Used in .EXT files.

Unknown

Fields

header_data: Vec<u8>

Unknown header data.

content_data: Vec<u8>

Unknown content data.

Unknown content.