rekordcrate::anlz
Enum ContentKind

pub enum ContentKind {

    File,
    BeatGrid,
    CueList,
    ExtendedCueList,
    ExtendedCue,
    Cue,
    Path,
    VBR,
    WaveformPreview,
    TinyWaveformPreview,
    WaveformDetail,
    WaveformColorPreview,
    WaveformColorDetail,
    SongStructure,
    Unknown([u8; 4]),

}

The kind of section.

Variants

File

File section that contains all other sections.

BeatGrid

All beats found in the track.

CueList

Either memory points and loops or hotcues and hot loops of the track.

Note: Since the release of the Nexus 2 series, there also exists the ExtendedCueList section which can carry additional information.

ExtendedCueList

Extended version of the CueList section (since Nexus 2 series).

ExtendedCue

Single cue entry inside a ExtendedCueList section.

Cue

Single cue entry inside a CueList section.

Path

File path of the audio file.

VBR

Seek information for variable bitrate files.

WaveformPreview

Fixed-width monochrome preview of the track waveform.

TinyWaveformPreview

Smaller version of the fixed-width monochrome preview of the track waveform (for the CDJ-900).

WaveformDetail

Variable-width large monochrome version of the track waveform.

Used in .EXT files.

WaveformColorPreview

Fixed-width colored version of the track waveform.

Used in .EXT files.

WaveformColorDetail

Variable-width large colored version of the track waveform.

Used in .EXT files.

SongStructure

Describes the structure of a sond (Intro, Chrous, Verse, etc.).

Used in .EXT files.

Unknown([u8; 4])

Unknown Kind.

This allows handling files that contain unknown section types and allows to access later sections in the file that have a known type instead of failing to parse the whole file.