rekordcrate::anlz
Struct Cue

pub struct Cue {
    pub header: Header,
    pub hot_cue: u32,
    pub status: u32,
    pub order_first: u16,
    pub order_last: u16,
    pub cue_type: CueType,
    pub time: u32,
    pub loop_time: u32,
    /* private fields */
}

A memory or hot cue (or loop).

Fields
header: Header

Cue entry header.

hot_cue: u32

Hot cue number.

Value	Hot cue

0	Not a hot cue.

1	A

2	B

…	…

status: u32

Loop status. 4 if this cue is an active loop, 0 otherwise.

order_first: u16

Somehow used for sorting cues.

Value	Cue

0xFFFF	1

0x0000	2

0x0002	3

0x0003	4

…	…

It is unknown why both order_first and order_last exist, when on of those values should suffice.

order_last: u16

Somehow used for sorting cues.

Value	Cue

0x0001	1

0x0002	2

0x0003	3

0x0004	4

…	…

0xFFFF	last

It is unknown why both order_first and order_last exist, when on of those values should suffice.

cue_type: CueType

Type of this cue (2 if this cue is a loop).

time: u32

Time in milliseconds after which this cue would occur (at normal playback speed).

loop_time: u32

Time in milliseconds after which this the loop would jump back to time (at normal playback speed).