Struct rekordcrate::anlz::ExtendedCue

pub struct ExtendedCue {

    pub header: Header,
    pub hot_cue: u32,
    pub cue_type: CueType,
    pub unknown1: u8,
    pub unknown2: u16,
    pub time: u32,
    pub loop_time: u32,
    pub color: ColorIndex,
    pub unknown3: u8,
    pub unknown4: u16,
    pub unknown5: u32,
    pub loop_numerator: u16,
    pub loop_denominator: u16,
    pub len_comment: u32,
    pub comment: String,
    pub hot_cue_color_index: u8,
    pub hot_cue_color_rgb: (u8, u8, u8),
    pub unknown6: u32,
    pub unknown7: u32,
    pub unknown8: u32,
    pub unknown9: u32,
    pub unknown10: u32,

}

A memory or hot cue (or loop).

Fields
header: Header

Cue entry header.

hot_cue: u32

Hot cue number.

Value	Hot cue

0	Not a hot cue.

1	A

2	B

…	…

cue_type: CueType

Type of this cue (2 if this cue is a loop).

unknown1: u8

Unknown field. Seems always have the value 0.

unknown2: u16

Unknown field. Seems always have the value 0x03E8 (= decimal 1000).

time: u32

Time in milliseconds after which this cue would occur (at normal playback speed).

loop_time: u32

Time in milliseconds after which this the loop would jump back to time (at normal playback speed).

color: ColorIndex

Color assigned to this cue.

Only used by memory cues, hot cues use a different value (see below).

unknown3: u8

Unknown field.

unknown4: u16

Unknown field.

unknown5: u32

Unknown field.

loop_numerator: u16

Represents the loop size numerator (if this is a quantized loop).

loop_denominator: u16

Represents the loop size denominator (if this is a quantized loop).

len_comment: u32

Length of the comment in bytes.

comment: String

And UTF-16BE encoded string, followed by a trailing 0x0000.

hot_cue_color_index: u8

Rekordbox hotcue color index.

| 0x00 | None (Green on older CDJs). | | 0x01 | #305aff | | 0x02 | #5073ff | | 0x03 | #508cff | | 0x04 | #50a0ff | | 0x05 | #50b4ff | | 0x06 | #50b0f2 | | 0x07 | #50aee8 | | 0x08 | #45acdb | | 0x09 | #00e0ff | | 0x0a | #19daf0 | | 0x0b | #32d2e6 | | 0x0c | #21b4b9 | | 0x0d | #20aaa0 | | 0x0e | #1fa392 | | 0x0f | #19a08c | | 0x10 | #14a584 | | 0x11 | #14aa7d | | 0x12 | #10b176 | | 0x13 | #30d26e | | 0x14 | #37de5a | | 0x15 | #3ceb50 | | 0x16 | #28e214 | | 0x17 | #7dc13d | | 0x18 | #8cc832 | | 0x19 | #9bd723 | | 0x1a | #a5e116 | | 0x1b | #a5dc0a | | 0x1c | #aad208 | | 0x1d | #b4c805 | | 0x1e | #b4be04 | | 0x1f | #bab404 | | 0x20 | #c3af04 | | 0x21 | #e1aa00 | | 0x22 | #ffa000 | | 0x23 | #ff9600 | | 0x24 | #ff8c00 | | 0x25 | #ff7500 | | 0x26 | #e0641b | | 0x27 | #e0461e | | 0x28 | #e0301e | | 0x29 | #e02823 | | 0x2a | #e62828 | | 0x2b | #ff376f | | 0x2c | #ff2d6f | | 0x2d | #ff127b | | 0x2e | #f51e8c | | 0x2f | #eb2da0 | | 0x30 | #e637b4 | | 0x31 | #de44cf | | 0x32 | #de448d | | 0x33 | #e630b4 | | 0x34 | #e619dc | | 0x35 | #e600ff | | 0x36 | #dc00ff | | 0x37 | #cc00ff | | 0x38 | #b432ff | | 0x39 | #b93cff | | 0x3a | #c542ff | | 0x3b | #aa5aff | | 0x3c | #aa72ff | | 0x3d | #8272ff | | 0x3e | #6473ff |

hot_cue_color_rgb: (u8, u8, u8)

Rekordbot hot cue color RGB value.

This color is similar but not identical to the color that Rekordbox displays, and possibly used to illuminate the RGB LEDs in a player that has loaded the cue. If not color is associated with this hot cue, the value is (0, 0, 0).

unknown6: u32

Unknown field.

unknown7: u32

Unknown field.

unknown8: u32

Unknown field.

unknown9: u32

Unknown field.

unknown10: u32

Unknown field.