Struct rekordcrate::anlz::WaveformColorDetailColumn

pub struct WaveformColorDetailColumn { /* private fields */ }

Single Column value in a Waveform Color Detail section.

Implementations

impl WaveformColorDetailColumn

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl WaveformColorDetailColumn

pub const fn into_bytes(self) -> [u8; 2]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 2]) -> Self

Converts the given bytes directly into the bitfield struct.

impl WaveformColorDetailColumn

pub fn red(&self) -> <B3 as Specifier>::InOut

Returns the value of red. Red color component.

pub fn red_or_err( &self ) -> Result<<B3 as Specifier>::InOut, InvalidBitPattern<<B3 as Specifier>::Bytes>>

Returns the value of red.

#Errors

If the returned value contains an invalid bit pattern for red. Red color component.

pub fn with_red(self, new_val: <B3 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of red set to the given value.

#Panics

If the given value is out of bounds for red. Red color component.

pub fn with_red_checked( self, new_val: <B3 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of red set to the given value.

#Errors

If the given value is out of bounds for red. Red color component.

pub fn set_red(&mut self, new_val: <B3 as Specifier>::InOut)

Sets the value of red to the given value.

#Panics

If the given value is out of bounds for red. Red color component.

pub fn set_red_checked( &mut self, new_val: <B3 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of red to the given value.

#Errors

If the given value is out of bounds for red. Red color component.

pub fn green(&self) -> <B3 as Specifier>::InOut

Returns the value of green. Green color component.

pub fn green_or_err( &self ) -> Result<<B3 as Specifier>::InOut, InvalidBitPattern<<B3 as Specifier>::Bytes>>

Returns the value of green.

#Errors

If the returned value contains an invalid bit pattern for green. Green color component.

pub fn with_green(self, new_val: <B3 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of green set to the given value.

#Panics

If the given value is out of bounds for green. Green color component.

pub fn with_green_checked( self, new_val: <B3 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of green set to the given value.

#Errors

If the given value is out of bounds for green. Green color component.

pub fn set_green(&mut self, new_val: <B3 as Specifier>::InOut)

Sets the value of green to the given value.

#Panics

If the given value is out of bounds for green. Green color component.

pub fn set_green_checked( &mut self, new_val: <B3 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of green to the given value.

#Errors

If the given value is out of bounds for green. Green color component.

pub fn blue(&self) -> <B3 as Specifier>::InOut

Returns the value of blue. Blue color component.

pub fn blue_or_err( &self ) -> Result<<B3 as Specifier>::InOut, InvalidBitPattern<<B3 as Specifier>::Bytes>>

Returns the value of blue.

#Errors

If the returned value contains an invalid bit pattern for blue. Blue color component.

pub fn with_blue(self, new_val: <B3 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of blue set to the given value.

#Panics

If the given value is out of bounds for blue. Blue color component.

pub fn with_blue_checked( self, new_val: <B3 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of blue set to the given value.

#Errors

If the given value is out of bounds for blue. Blue color component.

pub fn set_blue(&mut self, new_val: <B3 as Specifier>::InOut)

Sets the value of blue to the given value.

#Panics

If the given value is out of bounds for blue. Blue color component.

pub fn set_blue_checked( &mut self, new_val: <B3 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of blue to the given value.

#Errors

If the given value is out of bounds for blue. Blue color component.

pub fn height(&self) -> <B5 as Specifier>::InOut

Returns the value of height. Height of the column.

pub fn height_or_err( &self ) -> Result<<B5 as Specifier>::InOut, InvalidBitPattern<<B5 as Specifier>::Bytes>>

Returns the value of height.

#Errors

If the returned value contains an invalid bit pattern for height. Height of the column.

pub fn with_height(self, new_val: <B5 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of height set to the given value.

#Panics

If the given value is out of bounds for height. Height of the column.

pub fn with_height_checked( self, new_val: <B5 as Specifier>::InOut ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of height set to the given value.

#Errors

If the given value is out of bounds for height. Height of the column.

pub fn set_height(&mut self, new_val: <B5 as Specifier>::InOut)

Sets the value of height to the given value.

#Panics

If the given value is out of bounds for height. Height of the column.

pub fn set_height_checked( &mut self, new_val: <B5 as Specifier>::InOut ) -> Result<(), OutOfBounds>

Sets the value of height to the given value.

#Errors

If the given value is out of bounds for height. Height of the column.