Struct rekordcrate::anlz::WaveformPreviewColumn

pub struct WaveformPreviewColumn { /* fields omitted */ }

Single Column value in a Waveform Preview.

Implementations

impl WaveformPreviewColumn

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl WaveformPreviewColumn

pub const fn into_bytes(self) -> [u8; 1]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 1]) -> Self

Converts the given bytes directly into the bitfield struct.

impl WaveformPreviewColumn

pub fn height(&self) -> <B5 as Specifier>::InOut

Returns the value of height. Height of the Column in pixels.

pub fn height_or_err(
&self
) -> Result<<B5 as Specifier>::InOut, InvalidBitPattern<<B5 as Specifier>::Bytes>>

Returns the value of height.

#Errors

If the returned value contains an invalid bit pattern for height. Height of the Column in pixels.

pub fn with_height(self, new_val: <B5 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of height set to the given value.

#Panics

If the given value is out of bounds for height. Height of the Column in pixels.

pub fn with_height_checked(
self,
new_val: <B5 as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of height set to the given value.

#Errors

If the given value is out of bounds for height. Height of the Column in pixels.

pub fn set_height(&mut self, new_val: <B5 as Specifier>::InOut)

Sets the value of height to the given value.

#Panics

If the given value is out of bounds for height. Height of the Column in pixels.

pub fn set_height_checked(
&mut self,
new_val: <B5 as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of height to the given value.

#Errors

If the given value is out of bounds for height. Height of the Column in pixels.

pub fn whiteness(&self) -> <B3 as Specifier>::InOut

Returns the value of whiteness. Shade of white.

pub fn whiteness_or_err(
&self
) -> Result<<B3 as Specifier>::InOut, InvalidBitPattern<<B3 as Specifier>::Bytes>>

Returns the value of whiteness.

#Errors

If the returned value contains an invalid bit pattern for whiteness. Shade of white.

pub fn with_whiteness(self, new_val: <B3 as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of whiteness set to the given value.

#Panics

If the given value is out of bounds for whiteness. Shade of white.

pub fn with_whiteness_checked(
self,
new_val: <B3 as Specifier>::InOut
) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of whiteness set to the given value.

#Errors

If the given value is out of bounds for whiteness. Shade of white.

pub fn set_whiteness(&mut self, new_val: <B3 as Specifier>::InOut)

Sets the value of whiteness to the given value.

#Panics

If the given value is out of bounds for whiteness. Shade of white.

pub fn set_whiteness_checked(
&mut self,
new_val: <B3 as Specifier>::InOut
) -> Result<(), OutOfBounds>

Sets the value of whiteness to the given value.

#Errors

If the given value is out of bounds for whiteness. Shade of white.