rekordcrate::pdb::bitfields

Struct PageFlags

pub struct PageFlags { /* private fields */ }

Page flags stored in the page header.

Note that modular-bitfield stores the bits in LSB-first order so the bitfield definition is reversed compared to typical notation.

Implementations

impl PageFlags

pub const fn new() -> Self

Returns an instance with zero initialized data.

impl PageFlags

pub const fn into_bytes(self) -> [u8; 1]

Returns the underlying bits.

Layout

The returned byte array is layed out in the same way as described here.

pub const fn from_bytes(bytes: [u8; 1]) -> Self

Converts the given bytes directly into the bitfield struct.

impl PageFlags

pub fn unknown0(&self) -> <bool as Specifier>::InOut

Returns the value of unknown0. Unknown flag that appears to never be set.

pub fn unknown0_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown0.

Errors

If the returned value contains an invalid bit pattern for unknown0. Unknown flag that appears to never be set.

pub fn with_unknown0(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown0 set to the given value.

Panics

If the given value is out of bounds for unknown0. Unknown flag that appears to never be set.

pub fn with_unknown0_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown0 set to the given value.

Errors

If the given value is out of bounds for unknown0. Unknown flag that appears to never be set.

pub fn set_unknown0(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown0 to the given value.

Panics

If the given value is out of bounds for unknown0. Unknown flag that appears to never be set.

pub fn set_unknown0_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown0 to the given value.

Errors

If the given value is out of bounds for unknown0. Unknown flag that appears to never be set.

pub fn unknown1(&self) -> <bool as Specifier>::InOut

Returns the value of unknown1. Unknown flag that appears to never be set.

pub fn unknown1_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown1.

Errors

If the returned value contains an invalid bit pattern for unknown1. Unknown flag that appears to never be set.

pub fn with_unknown1(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown1 set to the given value.

Panics

If the given value is out of bounds for unknown1. Unknown flag that appears to never be set.

pub fn with_unknown1_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown1 set to the given value.

Errors

If the given value is out of bounds for unknown1. Unknown flag that appears to never be set.

pub fn set_unknown1(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown1 to the given value.

Panics

If the given value is out of bounds for unknown1. Unknown flag that appears to never be set.

pub fn set_unknown1_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown1 to the given value.

Errors

If the given value is out of bounds for unknown1. Unknown flag that appears to never be set.

pub fn unknown2(&self) -> <bool as Specifier>::InOut

Returns the value of unknown2. Unknown flag that appears to always be set.

pub fn unknown2_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown2.

Errors

If the returned value contains an invalid bit pattern for unknown2. Unknown flag that appears to always be set.

pub fn with_unknown2(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown2 set to the given value.

Panics

If the given value is out of bounds for unknown2. Unknown flag that appears to always be set.

pub fn with_unknown2_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown2 set to the given value.

Errors

If the given value is out of bounds for unknown2. Unknown flag that appears to always be set.

pub fn set_unknown2(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown2 to the given value.

Panics

If the given value is out of bounds for unknown2. Unknown flag that appears to always be set.

pub fn set_unknown2_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown2 to the given value.

Errors

If the given value is out of bounds for unknown2. Unknown flag that appears to always be set.

pub fn unknown3(&self) -> <bool as Specifier>::InOut

Returns the value of unknown3. Unknown flag that appears to never be set.

pub fn unknown3_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown3.

Errors

If the returned value contains an invalid bit pattern for unknown3. Unknown flag that appears to never be set.

pub fn with_unknown3(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown3 set to the given value.

Panics

If the given value is out of bounds for unknown3. Unknown flag that appears to never be set.

pub fn with_unknown3_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown3 set to the given value.

Errors

If the given value is out of bounds for unknown3. Unknown flag that appears to never be set.

pub fn set_unknown3(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown3 to the given value.

Panics

If the given value is out of bounds for unknown3. Unknown flag that appears to never be set.

pub fn set_unknown3_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown3 to the given value.

Errors

If the given value is out of bounds for unknown3. Unknown flag that appears to never be set.

pub fn contains_deleted(&self) -> <bool as Specifier>::InOut

Returns the value of contains_deleted. Flag set when the page contains a deleted row.

pub fn contains_deleted_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of contains_deleted.

Errors

If the returned value contains an invalid bit pattern for contains_deleted. Flag set when the page contains a deleted row.

pub fn with_contains_deleted(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of contains_deleted set to the given value.

Panics

If the given value is out of bounds for contains_deleted. Flag set when the page contains a deleted row.

pub fn with_contains_deleted_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of contains_deleted set to the given value.

Errors

If the given value is out of bounds for contains_deleted. Flag set when the page contains a deleted row.

pub fn set_contains_deleted(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of contains_deleted to the given value.

Panics

If the given value is out of bounds for contains_deleted. Flag set when the page contains a deleted row.

pub fn set_contains_deleted_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of contains_deleted to the given value.

Errors

If the given value is out of bounds for contains_deleted. Flag set when the page contains a deleted row.

pub fn unknown5(&self) -> <bool as Specifier>::InOut

Returns the value of unknown5. Unknown flag that appears to always be set.

pub fn unknown5_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown5.

Errors

If the returned value contains an invalid bit pattern for unknown5. Unknown flag that appears to always be set.

pub fn with_unknown5(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown5 set to the given value.

Panics

If the given value is out of bounds for unknown5. Unknown flag that appears to always be set.

pub fn with_unknown5_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown5 set to the given value.

Errors

If the given value is out of bounds for unknown5. Unknown flag that appears to always be set.

pub fn set_unknown5(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown5 to the given value.

Panics

If the given value is out of bounds for unknown5. Unknown flag that appears to always be set.

pub fn set_unknown5_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown5 to the given value.

Errors

If the given value is out of bounds for unknown5. Unknown flag that appears to always be set.

pub fn is_index_page(&self) -> <bool as Specifier>::InOut

Returns the value of is_index_page. Determines if the page is an index page.

pub fn is_index_page_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of is_index_page.

Errors

If the returned value contains an invalid bit pattern for is_index_page. Determines if the page is an index page.

pub fn with_is_index_page(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of is_index_page set to the given value.

Panics

If the given value is out of bounds for is_index_page. Determines if the page is an index page.

pub fn with_is_index_page_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of is_index_page set to the given value.

Errors

If the given value is out of bounds for is_index_page. Determines if the page is an index page.

pub fn set_is_index_page(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of is_index_page to the given value.

Panics

If the given value is out of bounds for is_index_page. Determines if the page is an index page.

pub fn set_is_index_page_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of is_index_page to the given value.

Errors

If the given value is out of bounds for is_index_page. Determines if the page is an index page.

pub fn unknown7(&self) -> <bool as Specifier>::InOut

Returns the value of unknown7. Unknown flag that appears to never be set.

pub fn unknown7_or_err( &self, ) -> Result<<bool as Specifier>::InOut, InvalidBitPattern<<bool as Specifier>::Bytes>>

Returns the value of unknown7.

Errors

If the returned value contains an invalid bit pattern for unknown7. Unknown flag that appears to never be set.

pub fn with_unknown7(self, new_val: <bool as Specifier>::InOut) -> Self

Returns a copy of the bitfield with the value of unknown7 set to the given value.

Panics

If the given value is out of bounds for unknown7. Unknown flag that appears to never be set.

pub fn with_unknown7_checked( self, new_val: <bool as Specifier>::InOut, ) -> Result<Self, OutOfBounds>

Returns a copy of the bitfield with the value of unknown7 set to the given value.

Errors

If the given value is out of bounds for unknown7. Unknown flag that appears to never be set.

pub fn set_unknown7(&mut self, new_val: <bool as Specifier>::InOut)

Sets the value of unknown7 to the given value.

Panics

If the given value is out of bounds for unknown7. Unknown flag that appears to never be set.

pub fn set_unknown7_checked( &mut self, new_val: <bool as Specifier>::InOut, ) -> Result<(), OutOfBounds>

Sets the value of unknown7 to the given value.

Errors

If the given value is out of bounds for unknown7. Unknown flag that appears to never be set.

impl PageFlags

pub fn new_data_page() -> Self

Create a PageFlags for a typical data page.

pub fn new_index_page() -> Self

Create a PageFlags for a typical index page.