Enum rekordcrate::pdb::Row

pub enum Row {

    Album {
        unknown1: u16,
        index_shift: u16,
        unknown2: u32,
        artist_id: u32,
        id: u32,
        unknown3: u32,
        unknown4: u8,
        name: DeviceSQLString,
    },
    Artist {
        subtype: u16,
        index_shift: u16,
        id: u32,
        unknown1: u8,
        ofs_name_near: u8,
        ofs_name_far: Option<u16>,
        name: DeviceSQLString,
    },
    Artwork {
        id: u32,
        path: DeviceSQLString,
    },
    Color {
        unknown1: u32,
        unknown2: u8,
        color: ColorIndex,
        unknown3: u16,
        name: DeviceSQLString,
    },
    Genre {
        id: u32,
        name: DeviceSQLString,
    },
    HistoryPlaylist {
        id: u32,
        name: DeviceSQLString,
    },
    HistoryEntry {
        track_id: u32,
        playlist_id: u32,
        entry_index: u32,
    },
    Key {
        id: u32,
        id2: u32,
        name: DeviceSQLString,
    },
    Label {
        id: u32,
        name: DeviceSQLString,
    },
    PlaylistTreeNode {
        parent_id: u32,
        unknown: u32,
        id: u32,
        sort_order: u32,
        node_is_folder: u32,
        name: DeviceSQLString,
    },
    PlaylistEntry {
        entry_index: u32,
        track_id: u32,
        playlist_id: u32,
    },
    Track {

        unknown1: u16,
        index_shift: u16,
        bitmask: u32,
        sample_rate: u32,
        composer_id: u32,
        file_size: u32,
        unknown2: u32,
        unknown3: u16,
        unknown4: u16,
        artwork_id: u32,
        key_id: u32,
        orig_artist_id: u32,
        label_id: u32,
        remixer_id: u32,
        bitrate: u32,
        track_number: u32,
        tempo: u32,
        genre_id: u32,
        album_id: u32,
        artist_id: u32,
        id: u32,
        disc_number: u16,
        play_count: u16,
        year: u16,
        sample_depth: u16,
        duration: u16,
        unknown5: u16,
        color: ColorIndex,
        rating: u8,
        unknown6: u16,
        unknown7: u16,
        isrc: DeviceSQLString,
        unknown_string1: DeviceSQLString,
        unknown_string2: DeviceSQLString,
        unknown_string3: DeviceSQLString,
        unknown_string4: DeviceSQLString,
        message: DeviceSQLString,
        kuvo_public: DeviceSQLString,
        autoload_hotcues: DeviceSQLString,
        unknown_string5: DeviceSQLString,
        unknown_string6: DeviceSQLString,
        date_added: DeviceSQLString,
        release_date: DeviceSQLString,
        mix_name: DeviceSQLString,
        unknown_string7: DeviceSQLString,
        analyze_path: DeviceSQLString,
        analyze_date: DeviceSQLString,
        comment: DeviceSQLString,
        title: DeviceSQLString,
        unknown_string8: DeviceSQLString,
        filename: DeviceSQLString,
        file_path: DeviceSQLString,
    
},
    Unknown,

}

A table row contains the actual data.

Variants

Album

Fields

unknown1: u16

Unknown field, usually 80 00.

index_shift: u16

Unknown field, called index_shift by @flesniak.

unknown2: u32

Unknown field.

artist_id: u32

ID of the artist row associated with this row.

id: u32

ID of this row.

unknown3: u32

Unknown field.

unknown4: u8

Unknown field.

name: DeviceSQLString

Byte offset of the album name string, relative to the start of this row.

Contains the album name, along with an ID of the corresponding artist.

Artist

Fields

subtype: u16

Determines if the name string is located at the 8-bit offset (0x60) or the 16-bit offset (0x64).

index_shift: u16

Unknown field, called index_shift by @flesniak.

id: u32

ID of this row.

unknown1: u8

Unknown field.

ofs_name_near: u8

One-byte name offset used if subtype is 0x60.

ofs_name_far: Option<u16>

Two-byte name offset used if subtype is 0x64.

In that case, the value of ofs_name_near is ignored

name: DeviceSQLString

Name of this artist.

Contains the artist name and ID.

Artwork

Fields

id: u32

ID of this row.

path: DeviceSQLString

Path to the album art file.

Contains the artwork path and ID.

Color

Fields

unknown1: u32

Unknown field.

unknown2: u8

Unknown field.

color: ColorIndex

Numeric color ID

unknown3: u16

Unknown field.

name: DeviceSQLString

User-defined name of the color.

Contains numeric color ID

Genre

Fields

id: u32

ID of this row.

name: DeviceSQLString

Name of the genre.

Represents a musical genre.

HistoryPlaylist

Fields

id: u32

ID of this row.

name: DeviceSQLString

Name of the playlist.

Represents a history playlist.

HistoryEntry

Fields

track_id: u32

ID of the track played at this position in the playlist.

playlist_id: u32

ID of the history playlist.

entry_index: u32

Position within the playlist.

Represents a history playlist.

Key

Fields

id: u32

ID of this row.

id2: u32

Apparently a second copy of the row ID.

name: DeviceSQLString

Name of the key.

Represents a musical key.

Label

Fields

id: u32

ID of this row.

name: DeviceSQLString

Name of the record label.

Represents a record label.

PlaylistTreeNode

Fields

parent_id: u32

ID of parent row of this row (which means that the parent is a folder).

unknown: u32

Unknown field.

id: u32

ID of this row.

sort_order: u32

Sort order indicastor.

node_is_folder: u32

Indicates if the node is a folder. Non-zero if it’s a leaf node, i.e. a playlist.

name: DeviceSQLString

Name of this node, as shown when navigating the menu.

Represents a node in the playlist tree (either a folder or a playlist).

PlaylistEntry

Fields

entry_index: u32

Position within the playlist.

track_id: u32

ID of the track played at this position in the playlist.

playlist_id: u32

ID of the history playlist.

Represents a track entry in a playlist.

Track

Fields

unknown1: u16

Unknown field, usually 24 00.

index_shift: u16

Unknown field, called index_shift by @flesniak.

bitmask: u32

Unknown field, called bitmask by @flesniak.

sample_rate: u32

Sample Rate in Hz.

composer_id: u32

Composer of this track as artist row ID (non-zero if set).

file_size: u32

File size in bytes.

unknown2: u32

Unknown field (maybe another ID?)

unknown3: u16

Unknown field (“always 19048?” according to @flesniak)

unknown4: u16

Unknown field (“always 30967?” according to @flesniak)

artwork_id: u32

Artwork row ID for the cover art (non-zero if set),

key_id: u32

Key row ID for the cover art (non-zero if set).

orig_artist_id: u32

Artist row ID of the original performer (non-zero if set).

label_id: u32

Label row ID of the original performer (non-zero if set).

remixer_id: u32

Artist row ID of the remixer (non-zero if set).

bitrate: u32

Bitrate of the track.

track_number: u32

Track number of the track.

tempo: u32

Track tempo in centi-BPM (= 1/100 BPM).

genre_id: u32

Genre row ID for this track (non-zero if set).

album_id: u32

Album row ID for this track (non-zero if set).

artist_id: u32

Artist row ID for this track (non-zero if set).

id: u32

Row ID of this track (non-zero if set).

disc_number: u16

Disc number of this track (non-zero if set).

play_count: u16

Number of times this track was played.

year: u16

Year this track was released.

sample_depth: u16

Bits per sample of the track aduio file.

duration: u16

Playback duration of this track in seconds (at normal speed).

unknown5: u16

Unknown field, apparently always “29”.

color: ColorIndex

Color row ID for this track (non-zero if set).

rating: u8

User rating of this track (0 to 5 starts).

unknown6: u16

Unknown field, apparently always “1”.

unknown7: u16

Unknown field (alternating “2” and “3”?).

isrc: DeviceSQLString

International Standard Recording Code (ISRC), in mangled format.

unknown_string1: DeviceSQLString

Unknown string field.

unknown_string2: DeviceSQLString

Unknown string field.

unknown_string3: DeviceSQLString

Unknown string field.

unknown_string4: DeviceSQLString

Unknown string field.

message: DeviceSQLString

Unknown string field (named by @flesniak).

kuvo_public: DeviceSQLString

Probably describes whether the track is public on kuvo.com (?). Value is either “ON” or empty string.

autoload_hotcues: DeviceSQLString

Determines if hotcues should be autoloaded. Value is either “ON” or empty string.

unknown_string5: DeviceSQLString

Unknown string field.

unknown_string6: DeviceSQLString

Unknown string field (usually empty).

date_added: DeviceSQLString

Date when the track was added to the Rekordbox collection.

release_date: DeviceSQLString

Date when the track was released.

mix_name: DeviceSQLString

Name of the remix (if any).

unknown_string7: DeviceSQLString

Unknown string field (usually empty).

analyze_path: DeviceSQLString

File path of the track analysis file.

analyze_date: DeviceSQLString

Date when the track analysis was performed.

comment: DeviceSQLString

Track comment.

title: DeviceSQLString

Track title.

unknown_string8: DeviceSQLString

Unknown string field (usually empty).

filename: DeviceSQLString

Name of the file.

file_path: DeviceSQLString

Path of the file.

Contains the album name, along with an ID of the corresponding artist.

Unknown

The row format (and also its size) is unknown, which means it can’t be parsed.