rekordcrate::pdb

Module ext

Parser for the extended Pioneer DeviceSQL database exports (exportExt.pdb).

The Rekordbox DJ software uses writes PDB files to /PIONEER/rekordbox/exportExt.pdb.

Most of the base file format has been reverse-engineered by Henry Betts, Fabian Lesniak and James Elliott. The exportExt specifics have been reversed engineered by Dominik Stolz (@voidc).

https://github.com/Deep-Symmetry/crate-digger/blob/master/doc/Analysis.pdf

https://djl-analysis.deepsymmetry.org/rekordbox-export-analysis/exports.html

https://github.com/henrybetts/Rekordbox-Decoding

https://github.com/flesniak/python-prodj-link/tree/master/prodj/pdblib

Structs

ParentId

A possibly absent parent ID. If the ID is zero (None), then there is no parent.

TagId

A unique identifier for a tag or category.

TagOrCategory

A tag or category that can be assigned to tracks for the purpose of categorization.

TagOrCategoryStrings

The strings associated with a tag or category.

TrackTag

M*N junction table between tags and tracks.

Enums

ExtPageType

The type of ext pages found inside a Table.

ExtRow

A table row contains the actual data.