rekordcrate::pdb

Module offset_array

Module for reading and writing offset arrays.

An offset array consists of an array of offsets, followed by data at those offsets. The offsets can be either u8 or u16, specified by the OffsetSize enum. The inner type T is constructed from an N-array of U, read from the positions at N offsets. The offsets are read/written in little-endian format. For example, an OffsetArrayContainer<T, 3> where T implements OffsetArrayItems will read/write 3 offsets, followed by 3 instances of T::Item at the offsets which are used to construct an inner T.

Example:

#[derive(Debug, PartialEq)]
struct SingleTarget<T: BinRead + BinWrite>(T);

impl<T: BinRead + BinWrite> OffsetArrayItems<1> for SingleTarget<T> {
    type Item = T;

    fn as_items(&self) -> [&Self::Item; 1] {
        [&self.0]
    }

    fn from_items(items: [Self::Item; 1]) -> Self {
     let [v] = items;
        Self(v)
    }
}

use rekordcrate::util::MaybeCalculated::*;
let near_u8_tail = OffsetArrayContainer {
    offsets: Provided([1u8].into()),
    inner: SingleTarget(42u8),
};

Structs

OffsetArrayContainer

A container for an array of offsets and the indirectly-addressed data at those offsets. See the module documentation for an example.

Enums

OffsetArray

The implementation of the offset array, which can be either u8 or u16. This is a private implementation detail, use OffsetArrayContainer instead. This enum is used to read/write the offsets, and to provide the read_offset and write_offset methods to read/write the inner type T at the specified offsets. The offsets are stored in little-endian format.

OffsetSize

Specifies whether the offsets are stored as u8 or u16.

Traits

OffsetArrayItems

Inner type that can be stored in an OffsetArrayContainer.