rekordcrate::pdb::offset_array

Struct OffsetArrayContainer

pub struct OffsetArrayContainer<T, const N: usize> {
    pub offsets: OffsetArray<N>,
    pub inner: T,
}

An array of offsets, followed by data at those offsets. The offsets are relative to the start of the offset array, minus an optional offset. The offsets can be either u8 or u16, specified by the OffsetSize The inner type T is read/written with the offsets and a base offset passed as arguments. The inner type T must implement BinRead and BinWrite, and its Args must be a tuple of (i64, &OffsetArray<N>, IA), where the first argument is the base offset, the second argument is a reference to the OffsetArray<N>, and the third argument is any additional arguments required by T. The OffsetArray itself takes as arguments a tuple of (usize, OffsetSize, IA), where the first argument is the offset to subtract from the start of the offset array to get the base offset, the second argument is the OffsetSize, and the third argument is any additional arguments required by T. The number of offsets is determined by the const generic parameter N. For example, an OffsetArray<T, 3> will read/write 3 offsets, followed by the data for T. The offsets are read/written in little-endian format. The inner type T is also read/written in little-endian format. This struct implements Deref to T, so the inner value can be accessed directly. Example:

#[binrw]
#[brw(little)]
#[br(import(base: i64, offsets: &OffsetArray<1>, args: <T as BinRead>::Args<'_>))]
#[bw(import(base: i64, offsets: &OffsetArray<1>, args: <T as BinWrite>::Args<'_>))]
#[derive(Debug, PartialEq)]
struct SingleTarget<T: BinRead + BinWrite>(
    #[brw(args(base, args))]
    #[br(parse_with = offsets.read_offset(0))]
    #[bw(write_with = offsets.write_offset(0))]
    T,
);

let near_u8_tail = OffsetArrayContainer {
    offsets: [1u8].into(),
    inner: SingleTarget(42u8),
};

Fields
offsets: OffsetArray<N>

The offsets, either u8 or u16.

inner: T

The inner value, read/written with the offsets.