Struct rekordcrate::pdb::string::DeviceSQLString

pub struct DeviceSQLString(_);

Encapsulates the intrinsics of the format used by DeviceSQL strings

Once A DeviceSQLString has been constructed, there is no way to change it.

use rekordcrate::pdb::string::DeviceSQLString;
use binrw::{BinWrite, BinRead};
let string = DeviceSQLString::new("foo".to_owned()).unwrap();
let binary = vec![0x9, 0x66, 0x6F, 0x6F];

let mut writer = binrw::io::Cursor::new(vec![]);
string.write(&mut writer)?;
assert_eq!(&binary, writer.get_ref());

let mut reader = binrw::io::Cursor::new(binary);
let parsed = DeviceSQLString::read(&mut reader)?;
assert_eq!(parsed, string);

Implementations

impl DeviceSQLString

pub fn new(string: String) -> Result<Self, StringError>

Initializes a DeviceSQLString from a plain Rust std::string::String

pub fn new_isrc(string: String) -> Result<Self, StringError>

Creates a DeviceSQLString containing an ISRC instead of an expected string

The DeviceSQL Database as used by rekordbox has a strange oddity where it serializes strings containing a tracks ISRC (International Standard Recording Code) in an unexpected format, if this is desired, use this constructor function instead of DeviceSQLString::new.

pub fn into_string(self) -> Result<String, StringError>

Extract the Rust string from the DeviceSQLString.

Consumes itself in the process.

pub const fn empty() -> Self

Create an empty DeviceSQLString.

Should be used to construct known empty strings.