rekordcrate::pdb

Trait RowVariant

pub trait RowVariant {
    const PAGE_TYPE: PageType;

    // Required methods
    fn from_row(row: &Row) -> Option<&Self>;
    fn from_row_mut(row: &mut Row) -> Option<&mut Self>;
}

A row variant that can be extracted from a generic Row.

Required Associated Constants

const PAGE_TYPE: PageType

The page type that contains rows of this variant.

Required Methods

fn from_row(row: &Row) -> Option<&Self>

Extracts a reference to this row variant from a generic Row.

fn from_row_mut(row: &mut Row) -> Option<&mut Self>

Extracts a mutable reference to this row variant from a generic Row.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.