Module rekordcrate::setting

Parser for Rekordbox *SETTING.DAT files.

These files either in the PIONEER directory of a USB drive (device exports), but are also present for on local installations of Rekordbox 6 in %APPDATA%\Pioneer\rekordbox6.

The settings files store the settings found on the “DJ System” > “My Settings” page of the Rekordbox preferences. These includes language, LCD brightness, tempo fader range, crossfader curve, etc.

The exact format still has to be reverse-engineered.

Structs

Setting

Represents a setting file.

Enums

AutoCue

Found at “PLAYER > DJ SETTING > AUTO CUE” of the “My Settings” page in the Rekordbox preferences.

AutoCueLevel

Found at “PLAYER > DJ SETTING > AUTO CUE LEVEL” of the “My Settings” page in the Rekordbox preferences.

BeatJumpBeatValue

Found at “PLAYER > DJ SETTING > BEAT JUMP BEAT VALUE” of the “My Settings” page in the Rekordbox preferences.

DiscSlotIllumination

Found at “PLAYER > DISPLAY(INDICATOR) > DISC SLOT ILLUMINATION” of the “My Settings” page in the Rekordbox preferences.

EjectLock

Found at “PLAYER > DJ SETTING > EJECT/LOAD LOCK” of the “My Settings” page in the Rekordbox preferences.

HotCueAutoLoad

Found at “PLAYER > DJ SETTING > HOT CUE AUTO LOAD” of the “My Settings” page in the Rekordbox preferences.

HotCueColor

Found at “PLAYER > DJ SETTING > HOT CUE COLOR” of the “My Settings” page in the Rekordbox preferences.

JogDisplayMode

Found at “PLAYER > DISPLAY(LCD) > JOG DISPLAY MODE” of the “My Settings” page in the Rekordbox preferences.

JogLCDBrightness

Found at “PLAYER > DISPLAY(LCD) > JOG LCD BRIGHTNESS” of the “My Settings” page in the Rekordbox preferences.

JogMode

Found at “PLAYER > DJ SETTING > JOG MODE” of the “My Settings” page in the Rekordbox preferences.

JogRingBrightness

Found at “PLAYER > DISPLAY(INDICATOR) > JOG RING BRIGHTNESS” of the “My Settings” page in the Rekordbox preferences.

JogRingIndicator

Found at “PLAYER > DISPLAY(INDICATOR) > JOG RING INDICATOR” of the “My Settings” page in the Rekordbox preferences.

LCDBrightness

Found at “PLAYER > DISPLAY(LCD) > LCD BRIGHTNESS” of the “My Settings” page in the Rekordbox preferences.

Language

Found at “PLAYER > DISPLAY(LCD) > LANGUAGE” of the “My Settings” page in the Rekordbox preferences.

MasterTempo

Found at “PLAYER > DJ SETTING > MASTER TEMPO” of the “My Settings” page in the Rekordbox preferences.

NeedleLock

Found at “PLAYER > DJ SETTING > NEEDLE LOCK” of the “My Settings” page in the Rekordbox preferences.

OnAirDisplay

Found at “PLAYER > DISPLAY(INDICATOR) > ON AIR DISPLAY” of the “My Settings” page in the Rekordbox preferences.

PadButtonBrightness

Found at “PLAYER > DISPLAY(INDICATOR) > PAD/BUTTON BRIGHTNESS” of the “My Settings” page in the Rekordbox preferences.

PhaseMeter

Found at “PLAYER > DJ SETTING > PHASE METER” of the “My Settings” page in the Rekordbox preferences.

PlayMode

Found at “PLAYER > DJ SETTING > PLAY MODE / AUTO PLAY MODE” of the “My Settings” page in the Rekordbox preferences.

Quantize

Found at “PLAYER > DJ SETTING > QUANTIZE” of the “My Settings” page in the Rekordbox preferences.

QuantizeBeatValue

Found at “PLAYER > DJ SETTING > QUANTIZE BEAT VALUE” of the “My Settings” page in the Rekordbox preferences.

SettingData

Data section of a *SETTING.DAT file.

SlipFlashing

Found at “PLAYER > DISPLAY(INDICATOR) > SLIP FLASHING” of the “My Settings” page in the Rekordbox preferences.

Sync

Found at “PLAYER > DJ SETTING > SYNC” of the “My Settings” page in the Rekordbox preferences.

TempoRange

Found at “PLAYER > DJ SETTING > TEMPO RANGE” of the “My Settings” page in the Rekordbox preferences.

TimeMode

Found at “PLAYER > DJ SETTING > TIME MODE” of the “My Settings” page in the Rekordbox preferences.

VinylSpeedAdjust

Found at “PLAYER > DJ SETTING > VINYL SPEED ADJUST” of the “My Settings” page in the Rekordbox preferences.

Waveform

Found at “PLAYER > DJ SETTING > WAVEFORM / PHASE METER” of the “My Settings” page in the Rekordbox preferences.

WaveformDivisions

Found at “PLAYER > DJ SETTING > WAVEFORM DIVISIONS” of the “My Settings” page in the Rekordbox preferences.