Crate rekordcrate

This library provides access to device libraries exported from Pioneer’s Rekordbox DJ software.

Re-exports

pub use crate::util::RekordcrateError as Error;

pub use crate::util::RekordcrateResult as Result;

Modules

anlz

Parser for Rekordbox analysis files, that can be found inside nested subdirectories of the PIONEER/USBANLZ directory and can have the extensions .DAT, .EXT or .2EX. Note that these files are not only used for device exports, but also for local rekordbox databases. In that case, the directory can be found at %APPDATA%\Pioneer\rekordbox\share\PIONEER\USBANLZ.

pdb

Parser for Pioneer DeviceSQL database exports (PDB).

setting

Parser for Rekordbox *SETTING.DAT files.

util

Common types used in multiple modules.