rekordcrate

Module anlz

Parser for Rekordbox analysis files, that can be found inside nested subdirectories of the PIONEER/USBANLZ directory and can have the extensions .DAT, .EXT or .2EX. Note that these files are not only used for device exports, but also for local rekordbox databases. In that case, the directory can be found at %APPDATA%\Pioneer\rekordbox\share\PIONEER\USBANLZ.

These files contain additional data (such as beatgrids, hotcues, waveforms and song structure information) that is not part of the PDB file.

The file is divided in section, where each section consists of a tag, header, and content.

With the evolution of the Pioneer hardware line, new section types were added (e.g. for high-resolution colored waveforms). To avoid issues with older hardware that cannot handle the additional data due to their memory limitations, the new sections were only added to a copy of the original file (.DAT) and saved with another extension (.EXT).

https://djl-analysis.deepsymmetry.org/rekordbox-export-analysis/anlz.html

https://reverseengineering.stackexchange.com/questions/4311/help-reversing-a-edb-database-file-for-pioneers-rekordbox-software

Structs

ANLZ

ANLZ file section.

Beat

A single beat inside the beat grid.

BeatGrid

All beats in the track.

Cue

A memory or hot cue (or loop).

CueList

List of cue points or loops (either hot cues or memory cues).

ExtendedCue

A memory or hot cue (or loop).

ExtendedCueList

List of cue points or loops (either hot cues or memory cues, extended version).

Header

Header of a section that contains type and size information.

Path

Path of the audio file that this analysis belongs to.

Phrase

A song structure entry that represents a phrase in the track.

Section

ANLZ Section.

SongStructure

Describes the structure of a song (Intro, Chrous, Verse, etc.).

SongStructureData

The data part of the SongStructure section that may be encrypted (RB6+).

TinyWaveformPreview

Smaller version of the fixed-width monochrome preview of the track waveform.

TinyWaveformPreviewColumn

Single Column value in a Tiny Waveform Preview.

Unknown

Unknown content.

VBR

Seek information for variable bitrate files (probably).

WaveformColorDetail

Variable-width large colored version of the track waveform.

WaveformColorDetailColumn

Single Column value in a Waveform Color Detail section.

WaveformColorPreview

Variable-width large monochrome version of the track waveform.

WaveformColorPreviewColumn

Single Column value in a Waveform Color Preview.

WaveformDetail

Variable-width large monochrome version of the track waveform.

WaveformPreview

Fixed-width monochrome preview of the track waveform.

WaveformPreviewColumn

Single Column value in a Waveform Preview.

Enums

Bank

Stylistic track bank for Lightning mode.

Content

Section content which differs depending on the section type.

ContentKind

The kind of section.

CueListType

Describes the types of entries found in a Cue List section.

CueType

Indicates if the cue is point or a loop.

Mood

Music classification that is used for Lightnight mode and based on rhythm, tempo kick drum and sound density.