Module rekordcrate::anlz

Parser for Rekordbox analysis files, that can be found inside nested subdirectories of the PIONEER/USBANLZ directory and can have the extensions .DAT, .EXT or .2EX.

These files contain additional data (such as beatgrids, hotcues, waveforms and song structure information) that is not part of the PDB file.

The file is divided in section, where each section consists of a tag, header, and content.

With the evolution of the Pioneer hardware line, new section types were added (e.g. for high-resolution colored waveforms). To avoid issues with older hardware that cannot handle the additional data due to their memory limitations, the new sections were only added to a copy of the original file (.DAT) and saved with another extension (.EXT).

https://djl-analysis.deepsymmetry.org/rekordbox-export-analysis/anlz.html

https://reverseengineering.stackexchange.com/questions/4311/help-reversing-a-edb-database-file-for-pioneers-rekordbox-software

Structs

ANLZ

ANLZ file section.

Beat

A single beat inside the beat grid.

Cue

A memory or hot cue (or loop).

ExtendedCue

A memory or hot cue (or loop).

Header

Header of a section that contains type and size information.

Phrase

A song structure entry that represents a phrase in the track.

Section

ANLZ Section.

TinyWaveformPreviewColumn

Single Column value in a Tiny Waveform Preview.

WaveformColorDetailColumn

Single Column value in a Waveform Color Detail section.

WaveformColorPreviewColumn

Single Column value in a Waveform Color Preview.

WaveformPreviewColumn

Single Column value in a Waveform Preview.

Enums

Bank

Stylistic track bank for Lightning mode.

Content

Section content which differs depending on the section type.

ContentKind

The kind of section.

CueListType

Describes the types of entries found in a Cue List section.

CueType

Indicates if the cue is point or a loop.

Mood

Music classification that is used for Lightnight mode and based on rhythm, tempo kick drum and sound density.