Module rekordcrate::pdb

Parser for Pioneer DeviceSQL database exports (PDB).

The Rekordbox DJ software uses writes PDB files to /PIONEER/rekordbox/export.pdb.

Most of the file format has been reverse-engineered by Henry Betts, Fabian Lesniak and James Elliott.

https://github.com/Deep-Symmetry/crate-digger/blob/master/doc/Analysis.pdf

https://djl-analysis.deepsymmetry.org/rekordbox-export-analysis/exports.html

https://github.com/henrybetts/Rekordbox-Decoding

https://github.com/flesniak/python-prodj-link/tree/master/prodj/pdblib

Structs

Header

The PDB header structure, including the list of tables.

Page

A table page.

PageIndex

Points to a table page and can be used to calculate the page’s file offset by multiplying it with the page size (found in the file header).

RowGroup

A group of row indices, which are built backwards from the end of the page. Holds up to sixteen row offsets, along with a bit mask that indicates whether each row is actually present in the table.

RowOffset

An offset which points to a row in the table, whose actual presence is controlled by one of the bits in row_present_flags. This instance allows the row itself to be lazily loaded, unless it is not present, in which case there is no content to be loaded.

Table

Tables are linked lists of pages containing rows of a single type, which are organized into groups.

Enums

DeviceSQLString

A string.

PageType

The type of pages found inside a Table.

Row

A table rows contains the actual data