Struct rekordcrate::pdb::Page

pub struct Page {
    pub page_index: PageIndex,
    pub page_type: PageType,
    pub next_page: PageIndex,
    pub num_rows_small: u8,
    pub page_flags: u8,
    pub free_size: u16,
    pub used_size: u16,
    pub num_rows_large: u16,
    // some fields omitted
}

A table page.

Each page consists of a header that contains information about the type, number of rows, etc., followed by the data section that holds the row data. Each row needs to be located using an offset found in the page footer at the end of the page.

Fields
page_index: PageIndex

Index of the page.

Should match the index used for lookup and can be used to verify that the correct page was loaded.

page_type: PageType

Type of information that the rows of this page contain.

Should match the page type of the table that this page belongs to.

next_page: PageIndex

Index of the next page with the same page type.

If this page is the last one of that type, the page index stored in the field will point past the end of the file.

num_rows_small: u8

Number of rows in this table (8-bit version).

Used if num_rows_large not greater than this value and not equal to 0x1FFF, which means that the number of rows fits into a single byte.

page_flags: u8

Page flags.

According to @flesniak:

strange pages: 0x44, 0x64; otherwise seen: 0x24, 0x34

free_size: u16

Free space in bytes in the data section of the page (excluding the row offsets in the page footer).

used_size: u16

Used space in bytes in the data section of the page.

num_rows_large: u16

Number of rows in this table (16-bit version).

Used when the number of rows does not fit into a single byte. In that case,num_rows_large is greater than num_rows_small, but is not equal to 0x1FFF.

Implementations

impl Page

pub fn has_data(&self) -> bool

Returns true if the page actually contains row data.

pub fn num_rows(&self) -> u16

Number of rows on this page.

Note that this number includes rows that have been flagged as missing by the row group.

pub fn num_row_groups(&self) -> u16

Number of row groups.

All row groups except the last one consist of 16 rows (but that number includes rows that have been flagged as missing by the row group.

pub fn row_groups<'a>(
&'a self,
page_data: &'a [u8],
page_size: u32
) -> impl Iterator<Item = RowGroup> + 'a

The rows groups found in this page.

pub fn row<'a>(
&self,
page_data: &'a [u8],
RowOffset: &RowOffset
) -> IResult<&'a [u8], Row>

Get the page row from the page_data slice.