rekordcrate::pdb

Struct Page

pub struct Page {
    pub page_index: PageIndex,
    pub page_type: PageType,
    pub next_page: PageIndex,
    pub num_rows_small: u8,
    pub free_size: u16,
    pub used_size: u16,
    pub content: PageContent,
    /* private fields */
}

A table page.

Each page consists of a header that contains information about the type, number of rows, etc., followed by the data section that holds the row data. Each row needs to be located using an offset found in the page footer at the end of the page.

Fields
page_index: PageIndex

Index of the page.

Should match the index used for lookup and can be used to verify that the correct page was loaded.

page_type: PageType

Type of information that the rows of this page contain.

Should match the page type of the table that this page belongs to.

next_page: PageIndex

Index of the next page with the same page type.

If this page is the last one of that type, the page index stored in the field will point past the end of the file.

num_rows_small: u8

Number of rows in this table (8-bit version).

Used if num_rows_large not greater than this value and not equal to 0x1FFF, which means that the number of rows fits into a single byte.

free_size: u16

Free space in bytes in the data section of the page (excluding the row offsets in the page footer).

used_size: u16

Used space in bytes in the data section of the page.

content: PageContent

The content of the page.

Implementations

impl Page

pub const HEADER_SIZE: u32 = 32u32

Size of the page header in bytes.