Struct rekordcrate::pdb::PageIndex

pub struct PageIndex(_);

Points to a table page and can be used to calculate the page’s file offset by multiplying it with the page size (found in the file header).

Implementations

impl PageIndex

pub fn offset(&self, page_size: u32) -> u64

Calculate the absolute file offset of the page in the PDB file for the given page_size.